pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "withenv"
authors = [{name = "Henning von Bargen", email = "hvbdev@von-bargen.eu"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "python-dotenv >= 1.0",
]

[project.scripts]
withenv = "withenv:main"